+++ packages/overlayscrollbars/test/playwright/setups/structureSetup/focus/handleEnvironment.ts
@@ -1,0 +1,24 @@
+import { addClass } from '~/support';
+
+{
+  const { body } = document;
+  const url = new URL(window.location.toString());
+  const params = url.searchParams;
+
+  /**
+   * nsh: native scrollbar hiding
+   * vpt: viewport is target
+   */
+  ['nsh', 'vpt'].forEach((param) => {
+    const paramValue = Boolean(params.get(param));
+
+    if (paramValue) {
+      addClass(body, param);
+    } else {
+      document.getElementById(param)?.addEventListener('click', () => {
+        params.set(param, 'true');
+        window.location.assign(url.toString());
+      });
+    }
+  });
+}